Add qemu-ifup, with dnsmasq
I guess I forgot it when I initially uploade the files.
Moved the command to start dnsmasq in host here. It does not work when it is in start.sh.

Look at previous commit to see why it is useful.
Also you may want to stop it in a qemu-ifdown script.