Change serial option to enable qemu monitor on stdio
When there is no serial port qemu console (monitor) is redirected to stdio.
But it does not work with serial. Worse, when you issue ctrl+c on terminal qemu is killed instead of the process inside vm.

This seems to be a proper way to restore default behaviour.
Taken from: http://unix.stackexchange.com/questions/167165/how-to-pass-ctrl-c-in-qemu